#47 セットリスト一覧の表示をコンポーネント化
