+++ app/javascript/components/Setlists.js
@@ -1,13 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom'
+import ListedSetlist from './ListedSetlist';
 
 const Setlists = () => {
-  const convertToHours = (milliSeconds) => {
-    const minutes = Math.ceil(milliSeconds / 1000) / 60
-    const hours = Math.trunc(minutes / 60)
-    const remainingMinutes = Math.ceil(minutes % 60)
-    return `${hours} 時間 ${remainingMinutes} 分`
-  }
+
 
     const [setlists, setSetlists] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
@@ -33,21 +29,23 @@ const Setlists = () => {
 
     return (
       <>
-        <h1 className="text-2xl text-center">セットリスト</h1>
-        <Link to="/setlists/new"><button>新規作成</button></Link>
-        <ul>
-          {setlists.map((setlist) => (
-            <li key={setlist.id}>
-              <Link to={`/setlists/${setlist.id}`}>
-                <div>{setlist.title}</div>
-                <span className="mr-2">{setlist.songs_count}曲</span>
-                <span>{convertToHours(setlist.total_duration_time)}</span>
-                <span className="mx-1">/</span>
-                <span>{convertToHours(setlist.target_duration_time)}</span>
-              </Link>
-            </li>
-          ))}
-        </ul>
+        <h1 className="text-2xl text-center mb-4 pt-4">セットリスト</h1>
+
+        <div className="p-4">
+          <ul>
+            <div className="mb-1 px-4 py-2 border">
+              <li>
+                <Link to="/setlists/new">
+                  <div>新規作成</div>
+                </Link>
+              </li>
+
+            </div>
+            {setlists.map((setlist) => (
+              <ListedSetlist key={setlist.id} value={setlist} />
+            ))}
+          </ul>
+        </div>
       </>
     )
 };
